Move things around for a more sensible setup responsibility
